+++ packages/video/src/index.ts
@@ -10,7 +10,7 @@ export { useVideo } from "./context";
 // Hooks
 export { useVolume } from "./hooks/use-volume";
 export { useTimeline } from "./hooks/use-timeline";
-export { useGetDuration } from "./hooks/get-duration";
+export { useGetDuration } from "./hooks/use-get-duration";
 export { usePlayPause } from "./hooks/use-play-pause";
 export { useMuteUnmute } from "./hooks/use-mute-unmute";
 export { useStartAt } from "./hooks/use-start-at";