+++ apps/cropper/CHANGELOG.md
@@ -1,5 +1,17 @@
 # web
 
+## 0.0.7
+
+### Patch Changes
+
+- @workspace/ui@0.0.2
+
+## 0.0.6
+
+### Patch Changes
+
+- @workspace/ui@0.0.1
+
 ## 0.0.5
 
 ### Patch Changes