+++ apps/test/CHANGELOG.md
@@ -1,5 +1,17 @@
 # test
 
+## 0.1.5
+
+### Patch Changes
+
+- @workspace/ui@0.0.2
+
+## 0.1.4
+
+### Patch Changes
+
+- @workspace/ui@0.0.1
+
 ## 0.1.3
 
 ### Patch Changes