+++ apps/video/CHANGELOG.md
@@ -1,5 +1,21 @@
 # web
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @zuude-ui/video@0.0.6
+  - @workspace/ui@0.0.2
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @zuude-ui/video@0.0.5
+  - @workspace/ui@0.0.1
+
 ## 0.0.5
 
 ### Patch Changes